add tomcat, support customization root

+++ build.sh
@@ -27,11 +27,11 @@ opt="$(getopt -o h -- "$@")" || usage "Parse options failed"
 
 eval set -- "${opt}"
 while true ; do
-    case "${1}" in
-    -h) usage ; shift ;;
-    --) shift ; break ;;
-    *) echo "Internal error!" ; exit 1 ;;
-    esac
+	case "${1}" in
+	-h) usage ; shift ;;
+	--) shift ; break ;;
+	*) echo "Internal error!" ; exit 1 ;;
+	esac
 done
 
 function check_copy_file() {
@@ -47,6 +47,7 @@ function check_copy_file() {
 		fi
 	done
 
+	echo "no file found for '${check_path}'"
 	return 1
 }
 
@@ -59,50 +60,17 @@ function build() {
 		cp -a "${PD}/${imgname}/${tag}" "${PD}/tmp/${imgname}/" || exit $?
 		sed -i "s/DOCKER_BASE/${DOCKER_BASE}/g" "${PD}/tmp/${imgname}/${tag}/Dockerfile" || exit $?
 		pushd "${PD}/tmp/${imgname}/${tag}" >/dev/null || exit $?
-		case "${imgname}" in
-		ubuntu)
-			case "${tag}" in
-			12.04)
-				docker build -t ${DOCKER_BASE}/${imgname} -rm . || exit $?
-				docker tag ${DOCKER_BASE}/${imgname} ${DOCKER_BASE}/${imgname}:${tag} || exit $?
-				;;
-			dev)
-				docker build -t ${DOCKER_BASE}/${imgname}:${tag} -rm . || exit $?
-				;;
-			esac
-			;;
-		golang)
-			case "${tag}" in
-			1.2)
-				if ! check_copy_file "go1.2.linux-amd64.tar.gz" "../go1.2.linux-amd64.tar.gz" "../dev/go1.2.linux-amd64.tar.gz" ; then
-					wget "https://go.googlecode.com/files/go1.2.linux-amd64.tar.gz" || exit $?
-				fi
-				sha1sum -c sha1sum || exit $?
-				docker build -t ${DOCKER_BASE}/${imgname} -rm . || exit $?
-				docker tag ${DOCKER_BASE}/${imgname} ${DOCKER_BASE}/${imgname}:${tag} || exit $?
-				;;
-			dev)
-				check_copy_file "build.sh" "../../ubuntu/dev/build.sh"  || exit $?
-				docker build -t ${DOCKER_BASE}/${imgname}:${tag} -rm . || exit $?
-				;;
-			esac
-			;;
-		nginx)
-			case "${tag}" in
-			latest)
-				docker build -t ${DOCKER_BASE}/${imgname} -rm . || exit $?
-				;;
-			dev)
-				check_copy_file "build.sh" "../../ubuntu/dev/build.sh"  || exit $?
-				docker build -t ${DOCKER_BASE}/${imgname}:${tag} -rm . || exit $?
-				;;
-			esac
-			;;
-		*)
-			docker build -t ${DOCKER_BASE}/${imgname} -rm . || exit $?
-			docker tag ${DOCKER_BASE}/${imgname} ${DOCKER_BASE}/${imgname}:${tag} || exit $?
-			;;
-		esac
+		check_copy_file "build-post.sh" "../../../ubuntu/build-post.sh" || exit $?
+		if [ "${tag}" == "dev" ] ; then
+			check_copy_file "build.sh" "../../../ubuntu/dev/build.sh" || exit $?
+		fi
+		if [ "${imgname}" == "golang" ] && [ "${tag}" == "1.2" ] ; then
+			if ! check_copy_file "go1.2.linux-amd64.tar.gz" "../go1.2.linux-amd64.tar.gz" ; then
+				wget "https://go.googlecode.com/files/go1.2.linux-amd64.tar.gz" || exit $?
+			fi
+			sha1sum -c sha1sum || exit $?
+		fi
+		docker build -t ${DOCKER_BASE}/${imgname}:${tag} -rm . || exit $?
 		popd >/dev/null || exit $?
 	done
 }
@@ -112,7 +80,7 @@ if [ $# -eq 0 ] ; then
 	build golang 1.2 dev
 	build nginx latest dev
 	build java jre7
-	build tomcat 7
+	build tomcat 7 dev
 else
 	for i in "$@" ; do
 		case "${i}" in
@@ -120,7 +88,7 @@ else
 		golang) build ${i} 1.2 dev ;;
 		nginx) build ${i} latest dev ;;
 		java) build ${i} jre7 ;;
-		tomcat) build ${i} 7 ;;
+		tomcat) build ${i} 7 dev ;;
 		esac
 	done
 fi
